+++ packages/core/solidity/src/index.ts
@@ -19,6 +19,6 @@ export { OptionsError } from './error';
 export type { Kind } from './kind';
 export { sanitizeKind } from './kind';
 
-export { erc20, erc721, erc1155, stablecoin, realWorldAsset, account, paymaster, governor, custom } from './api';
+export { erc20, erc721, erc1155, stablecoin, realWorldAsset, account, governor, custom } from './api';
 
 export { compatibleContractsSemver } from './utils/version';